requirements.txt
gymnasium==0.27.1
gymnasium_robotics==1.2.0
mujoco==2.3.1.post1
numpy==1.24.1
Pillow==9.4.0
setuptools==67.1.0
